Create 392.IsSubsequenceLeetcode12.py
# q link  https://leetcode.com/problems/is-subsequence/
392. Is Subsequence
Easy
6.5K
365
Companies
Given two strings s and t, return true if s is a subsequence of t, or false otherwise.

A subsequence of a string is a new string that is formed from the original string by deleting some (can be none) of the characters without disturbing the relative positions of the remaining characters. (i.e., "ace" is a subsequence of "abcde" while "aec" is not).